back: setup postsController.create()

+++ back/src/main/java/com/openclassrooms/mddapi/repository/TopicRepository.java
@@ -3,6 +3,6 @@ package com.openclassrooms.mddapi.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.openclassrooms.mddapi.model.Topic;
 
-public interface TopicRepository extends JpaRepository<Topic, Integer> {
+public interface TopicRepository extends JpaRepository<Topic, Long> {
 
 }
